refactor: move modifiers implementations in addon folder
